1.1.5

+++ src/modules/AnnotationModule.ts
@@ -96,7 +96,7 @@ export default class AnnotationModule implements ReaderModule {
 
         this.delegate.annotationModule = this
 
-        this.highlightsView = HTMLUtilities.findElement(this.headerMenu, "#container-view-highlights") as HTMLDivElement;
+        if (this.headerMenu) this.highlightsView = HTMLUtilities.findElement(this.headerMenu, "#container-view-highlights") as HTMLDivElement;
 
         if (this.initialAnnotations) {
             var highlights = this.initialAnnotations['highlights'] || null;
@@ -163,9 +163,16 @@ export default class AnnotationModule implements ReaderModule {
         }
     }
 
+    public async deleteAnnotation(highlight: Annotation): Promise<any> {
+        this.deleteLocalHighlight(highlight.id);
+    }
+    public async addAnnotation(highlight: Annotation): Promise<any> {
+        await this.annotator.saveAnnotation(highlight);
+        await this.drawHighlights();
+    }
 
-    private async deleteHighlight(highlight: Annotation): Promise<any> {
-        if (this.api) {
+    public async deleteHighlight(highlight: Annotation): Promise<any> {
+        if (this.api && this.api.deleteAnnotation) {
             this.api.deleteAnnotation(highlight).then(async () => {
                 this.deleteLocalHighlight(highlight.id);
             })
@@ -175,7 +182,7 @@ export default class AnnotationModule implements ReaderModule {
     }
 
     public async deleteSelectedHighlight(highlight: Annotation): Promise<any> {
-        if (this.api) {
+        if (this.api && this.api.deleteAnnotation) {
             this.api.deleteAnnotation(highlight).then(async () => {
                 this.deleteLocalHighlight(highlight.id);
             })
@@ -220,7 +227,7 @@ export default class AnnotationModule implements ReaderModule {
                     hightlight: highlight.selectionInfo.cleanText
                 }
             }
-            if (this.api) {
+            if (this.api && this.api.addAnnotation) {
                 this.api.addAnnotation(annotation).then(async result => {
                     annotation.id = result.id
                     var saved = await this.annotator.saveAnnotation(annotation);
